+++ src/Contracts/HttpStatusExceptionInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Codemonster\Support\Contracts;
+
+interface HttpStatusExceptionInterface extends \Throwable
+{
+    public function getStatusCode(): int;
+}